--- fraud/src/main/java/com/bartoszmarkiewicz/fraud/FraudRepository.java
@@ -1,8 +1,0 @@
-package com.bartoszmarkiewicz.fraud;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FraudRepository extends JpaRepository<Fraud, Integer> {
-}
