+++ inventory/src/main/java/com/bartoszmarkiewicz/inventory/repository/ProductRepository.java
@@ -1,0 +1,22 @@
+package com.bartoszmarkiewicz.inventory.repository;
+
+import com.bartoszmarkiewicz.inventory.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+    @Query("" +
+            "SELECT CASE WHEN COUNT(p) > 0 THEN " +
+            "TRUE ELSE FALSE END " +
+            "FROM Product p " +
+            "WHERE p.productName = ?1"
+    )
+    Boolean selectExistsProduct(String productName);
+
+
+
+
+}
