--- notification/src/main/java/com/bartoszmarkiewicz/notification/NotificationConfig.java
@@ -1,44 +1,0 @@
-package com.bartoszmarkiewicz.notification;
-
-import lombok.Getter;
-import org.springframework.amqp.core.Binding;
-import org.springframework.amqp.core.BindingBuilder;
-import org.springframework.amqp.core.Queue;
-import org.springframework.amqp.core.TopicExchange;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@Getter
-public class NotificationConfig {
-
-    @Value("${rabbitmq.exchanges.internal}")
-    private String internalExchange;
-
-    @Getter
-    @Value("${rabbitmq.queues.notification}")
-    private String notificationQueue;
-
-    @Value("${rabbitmq.routing-keys.internal-notification}")
-    private String internalNotificationRoutingKey;
-
-    @Bean
-    public TopicExchange internalTopicExchange() {
-        return new TopicExchange(this.internalExchange);
-    }
-
-    @Bean
-    public Queue notificationQueue() {
-        return new Queue(this.notificationQueue);
-    }
-
-    @Bean
-    public Binding internalToNotificationBinding() {
-        return BindingBuilder
-                .bind(notificationQueue())
-                .to(internalTopicExchange())
-                .with(this.internalNotificationRoutingKey);
-    }
-
-}
