--- notification/src/main/java/com/bartoszmarkiewicz/notification/NotificationController.java
@@ -1,24 +1,0 @@
-package com.bartoszmarkiewicz.notification;
-
-import com.bartoszmarkiewicz.clients.notification.NotificationRequest;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("api/v1/notification")
-@AllArgsConstructor
-@Slf4j
-public class NotificationController {
-
-    private NotificationService notificationService;
-
-    @PostMapping
-    public void sendNotification(@RequestBody NotificationRequest notificationRequest) {
-        log.info("New notification... {}", notificationRequest);
-        notificationService.send(notificationRequest);
-    }
-}
