--- notification/src/main/java/com/bartoszmarkiewicz/notification/RepositoryNotification.java
@@ -1,6 +1,0 @@
-package com.bartoszmarkiewicz.notification;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RepositoryNotification extends JpaRepository<Notification, Integer> {
-}
