+++ notification/src/main/java/com/bartoszmarkiewicz/notification/controller/NotificationController.java
@@ -1,0 +1,25 @@
+package com.bartoszmarkiewicz.notification.controller;
+
+import com.bartoszmarkiewicz.clients.notification.NotificationRequest;
+import com.bartoszmarkiewicz.notification.service.NotificationService;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("api/v1/notification")
+@AllArgsConstructor
+@Slf4j
+public class NotificationController {
+
+    private NotificationService notificationService;
+
+    @PostMapping
+    public void sendNotification(@RequestBody NotificationRequest notificationRequest) {
+        log.info("New notification... {}", notificationRequest);
+        notificationService.send(notificationRequest);
+    }
+}
