+++ notification/src/main/java/com/bartoszmarkiewicz/notification/repository/RepositoryNotification.java
@@ -1,0 +1,7 @@
+package com.bartoszmarkiewicz.notification.repository;
+
+import com.bartoszmarkiewicz.notification.model.Notification;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RepositoryNotification extends JpaRepository<Notification, Integer> {
+}
