--- order/src/main/java/com/bartoszmarkiewicz/order/OrderRegistrationRequest.java
@@ -1,20 +1,0 @@
-package com.bartoszmarkiewicz.order;
-
-import java.time.LocalDateTime;
-
-
-public record OrderRegistrationRequest(
-        Integer orderId,
-        Integer customerId,
-
-        Integer productId,
-
-        String productName,
-        Double orderValue,
-        Integer orderAmount,
-        LocalDateTime orderCreatedAt,
-        String shippingAddress,
-        String phoneNumber,
-        OrderStatus orderStatus) {
-
-}
