--- order/src/main/java/com/bartoszmarkiewicz/order/OrderRepository.java
@@ -1,8 +1,0 @@
-package com.bartoszmarkiewicz.order;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-}
