--- order/src/main/java/com/bartoszmarkiewicz/order/OrderStatus.java
@@ -1,7 +1,0 @@
-package com.bartoszmarkiewicz.order;
-
-public enum OrderStatus {
-    PROGRESS,
-    FINISHED
-
-}
