+++ order/src/main/java/com/bartoszmarkiewicz/order/config/OrderConfig.java
@@ -1,0 +1,18 @@
+package com.bartoszmarkiewicz.order.config;
+
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class OrderConfig {
+
+    @Bean
+    @LoadBalanced  //Load Balancer
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+
+
+}
