+++ order/src/main/java/com/bartoszmarkiewicz/order/model/OrderStatus.java
@@ -1,0 +1,7 @@
+package com.bartoszmarkiewicz.order.model;
+
+public enum OrderStatus {
+    PROGRESS,
+    FINISHED
+
+}
