+++ order/src/main/java/com/bartoszmarkiewicz/order/repository/OrderRepository.java
@@ -1,0 +1,9 @@
+package com.bartoszmarkiewicz.order.repository;
+
+import com.bartoszmarkiewicz.order.model.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+}
