+++ order/src/main/resources/application.yml
@@ -7,6 +7,7 @@ spring:
     username: postgres
     password: admin
     url: jdbc:postgresql://localhost:5432/order
+
   jpa:
     hibernate:
       ddl-auto: create-drop
