+++ order/src/test/resources/config/application.yml
@@ -1,0 +1,5 @@
+spring:
+  jpa:
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.H2Dialect
